requirements.txt
tinker
tinker-cookbook
playwright
Pillow
scikit-image
numpy
torch
tqdm
transformers
datasets
open_clip_torch
